this is the first commit
Flutter Todos is a simple Flutter app that allows you to manage your to-do list. With Flutter Todos, you can easily add, edit, and delete tasks as needed, and mark them as complete as you finish them. This app is great for anyone who wants to stay organized and keep track of their daily tasks.

Features
Add new to-do items
Edit existing to-do items
Delete to-do items
Mark to-do items as complete
Swipe to delete to-do items
Clean and intuitive user interface
Installation
To use Flutter Todos, simply clone this repository and run the app in your Flutter development environment.

Usage
To use Flutter Todos, simply open the app and start adding your to-do items. To add a new item, click on the floating action button in the bottom right-hand corner of the screen. To edit an item, simply tap on it and make your changes. To mark an item as complete, simply check the checkbox next to it. To delete an item, swipe left or right on the item and click the delete icon that appears.

Contributing
If you would like to contribute to Flutter Todos, please feel free to fork this repository and submit a pull request with your changes. Contributions are always welcome!

License
Flutter Todos is open source software licensed under the MIT License.